requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
docopt==0.6.2
easyocr==1.7.0
filelock==3.12.0
idna==3.4
imageio==2.29.0
Jinja2==3.1.2
jsonpickle==3.0.1
lazy_loader==0.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.24.3
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.1
pafy==0.5.5
Pillow==9.5.0
pipreqs==0.4.13
pyclipper==1.3.0.post4
python-bidi==0.4.2
PyWavelets==1.4.1
PyYAML==6.0
requests==2.31.0
scikit-image==0.20.0
scipy==1.10.1
shapely==2.0.1
six==1.16.0
sympy==1.12
tifffile==2023.4.12
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
typing_extensions==4.6.2
urllib3==2.0.2
yarg==0.1.9
youtube-dl==2021.12.17
